+++ src/app/stillinger/(sok)/elastic/filter/filterPublished.ts
@@ -1,4 +1,6 @@
-export function filterPublished(published: string | undefined) {
+import type { RangeFilter } from "@/app/stillinger/(sok)/elastic/types/types";
+
+export function filterPublished(published: string | undefined): RangeFilter[] {
     const filters = [];
     if (published) {
         filters.push({
